Oh noez, xfdesktop does not work in Natty! Squashed more small bugs, leaving only LibreOffice image because the system is useless otherwise

+++ build.sh
@@ -5,12 +5,12 @@ fail() { echo "Compilation failed!" ; exit 1; }
 cd fakechroot
 [ -e configure ] || ./autogen.sh || fail
 [ -e config.h ] || ./configure --host=arm-linux-gnueabi --prefix=/usr || fail
-make -j4 && cp -f src/.libs/libfakechroot.so .. && arm-linux-gnueabi-strip -g ../libfakechroot.so || fail
+[ -e ../libfakechroot.so ] || { make -j4 && cp -f src/.libs/libfakechroot.so .. && arm-linux-gnueabi-strip -g ../libfakechroot.so || fail ; }
 cd ..
 
 cd c-ares
 [ -e ares_config.h ] || ./configure --host=arm-linux-gnueabi --prefix=/usr || fail
-make -j4 && cp -f .libs/libcares.so ../libfakedns.so && arm-linux-gnueabi-strip -g ../libfakedns.so || fail
+[ -e ../libfakedns.so ] || { make -j4 && cp -f .libs/libcares.so ../libfakedns.so && arm-linux-gnueabi-strip -g ../libfakedns.so || fail ; }
 cd ..
 
 cd androidVNC/ZoomerWithKeys
